Fix workflow-guard-tests: move Settings-open entrypoints out of AppDelegate.swift
AppDelegate.swift is over the 900-line hard cap, so it may not grow at
all in a PR; the presentPreferencesWindow failure-surfacing change added
5 net lines. Move presentPreferencesWindow/openPreferencesWindow into
Sources/App/AppDelegateSettingsPresentation.swift (AppDelegate now
shrinks by 29 lines net).
